+++ mlens/utils/tests/test_utils.py
@@ -54,6 +54,8 @@ def test_print_time():
     # Record recorded print_time message
     utils.print_time(t0, message='test', file=logger)
 
+    print(logger.log[0])
+
     assert logger.log[0][:15] == 'test | 00:00:01'
